+++ packages/dynamic-bonding-curve/CHANGELOG.md
@@ -73,6 +73,7 @@
 - feat: added `tempWSolAcc` param to `claimCreatorTradingFee` and `claimPartnerTradingFee` functions
 - feat: added `calculateInitialPriceFromSqrtStartPrice`, `calculateFeeScheduler` and `calculateLockedVesting` functions
 - fix: reduced client-side filtering in `getPoolsQuoteFeesByConfig` and `getPoolsBaseFeesByConfig`
+- feat: removed `getTokenDecimal` state function from client.state
 
 #### Breaking Changes